+++ python/src/aiconfig/editor/client/src/components/AIConfigEditor.tsx
@@ -155,7 +155,7 @@ export type AIConfigCallbacks = {
 type RequestCallbackError = { message?: string };
 
 function AIConfigEditorBase({
-  aiconfig: initialAIConfig,
+  aiconfig: providedAIConfig,
   callbacks,
   mode,
   readOnly = false,
@@ -164,8 +164,20 @@ function AIConfigEditorBase({
   const [serverStatus, setServerStatus] = useState<"OK" | "ERROR">("OK");
   const [aiconfigState, dispatch] = useReducer(
     aiconfigReducer,
-    aiConfigToClientConfig(initialAIConfig)
+    aiConfigToClientConfig(providedAIConfig)
   );
+
+  const [prevProvidedConfig, setPrevProvidedConfig] =
+    useState<AIConfig>(providedAIConfig);
+  // After initializing the aiconfigState, we should also support updating the
+  // state if the provided AIConfig changes externally
+  if (prevProvidedConfig !== providedAIConfig) {
+    setPrevProvidedConfig(providedAIConfig);
+    dispatch({
+      type: "PROVIDED_AICONFIG_UPDATE",
+      config: providedAIConfig,
+    });
+  }
 
   const { showNotification } = useContext(NotificationContext);
 
@@ -182,6 +194,7 @@ function AIConfigEditorBase({
     }
     try {
       await downloadCallback();
+      logEventHandler?.("DOWNLOAD_BUTTON_CLICKED");
     } catch (err: unknown) {
       const message = (err as RequestCallbackError).message ?? null;
       showNotification({
@@ -190,7 +203,7 @@ function AIConfigEditorBase({
         type: "error",
       });
     }
-  }, [downloadCallback, showNotification]);
+  }, [downloadCallback, logEventHandler, showNotification]);
 
   const shareCallback = callbacks?.share;
   const onShare = useCallback(async () => {
@@ -199,6 +212,7 @@ function AIConfigEditorBase({
     }
     try {
       const { share_url: shareUrl } = await shareCallback();
+      logEventHandler?.("SHARE_BUTTON_CLICKED");
       return shareUrl;
     } catch (err: unknown) {
       const message = (err as RequestCallbackError).message ?? null;
@@ -208,7 +222,7 @@ function AIConfigEditorBase({
         type: "error",
       });
     }
-  }, [shareCallback, showNotification]);
+  }, [logEventHandler, shareCallback, showNotification]);
 
   const saveCallback = callbacks?.save;
   const onSave = useCallback(async () => {
@@ -242,8 +256,10 @@ function AIConfigEditorBase({
       async (
         promptName: string,
         newPrompt: Prompt,
-        onSuccess: (aiconfigRes: AIConfig) => void,
-        onError: (err: unknown) => void
+        callbacks: {
+          onSuccess?: (aiconfigRes: AIConfig) => void;
+          onError?: (err: unknown) => void;
+        }
       ) => {
         try {
           const serverConfigRes = await updatePromptCallback(
@@ -251,10 +267,10 @@ function AIConfigEditorBase({
             newPrompt
           );
           if (serverConfigRes?.aiconfig) {
-            onSuccess(serverConfigRes.aiconfig);
+            callbacks?.onSuccess?.(serverConfigRes.aiconfig);
           }
         } catch (err: unknown) {
-          onError(err);
+          callbacks?.onError?.(err);
         }
       },
       DEBOUNCE_MS
@@ -276,6 +292,7 @@ function AIConfigEditorBase({
       };
 
       dispatch(action);
+      logEventHandler?.("UPDATE_PROMPT_INPUT");
 
       const onError = (err: unknown) => {
         const message = (err as RequestCallbackError).message ?? null;
@@ -299,19 +316,21 @@ function AIConfigEditorBase({
             ...prompt,
             input: newPromptInput,
           },
-          (config) =>
-            dispatch({
-              type: "CONSOLIDATE_AICONFIG",
-              action,
-              config,
-            }),
-          onError
+          {
+            onSuccess: (config) =>
+              dispatch({
+                type: "CONSOLIDATE_AICONFIG",
+                action,
+                config,
+              }),
+            onError,
+          }
         );
       } catch (err: unknown) {
         onError(err);
       }
     },
-    [debouncedUpdatePrompt, dispatch, showNotification]
+    [debouncedUpdatePrompt, dispatch, logEventHandler, showNotification]
   );
 
   const onChangePromptName = useCallback(
@@ -347,19 +366,23 @@ function AIConfigEditorBase({
           // PromptName component maintains local state for the name to show in the UI
           // We cannot update client config state until the name is successfully set server-side
           // or else we could end up referencing a prompt name that is not set server-side
-          () =>
-            dispatch({
-              type: "UPDATE_PROMPT_NAME",
-              id: promptId,
-              name: newName,
-            }),
-          onError
+          {
+            onSuccess: () => {
+              dispatch({
+                type: "UPDATE_PROMPT_NAME",
+                id: promptId,
+                name: newName,
+              });
+              logEventHandler?.("UPDATE_PROMPT_NAME");
+            },
+            onError,
+          }
         );
       } catch (err: unknown) {
         onError(err);
       }
     },
-    [debouncedUpdatePrompt, showNotification]
+    [debouncedUpdatePrompt, logEventHandler, showNotification]
   );
 
   const updateModelCallback = callbacks?.updateModel;
@@ -387,6 +410,57 @@ function AIConfigEditorBase({
     );
   }, [updateModelCallback]);
 
+  const onUpdatePromptMetadata = useCallback(
+    async (promptId: string, newMetadata: JSONObject) => {
+      if (!debouncedUpdatePrompt) {
+        // Just no-op if no callback specified. We could technically perform
+        // client-side updates but that might be confusing
+        return;
+      }
+
+      dispatch({
+        type: "UPDATE_PROMPT_METADATA",
+        id: promptId,
+        metadata: newMetadata,
+      });
+
+      const onError = (err: unknown) => {
+        const message = (err as RequestCallbackError).message ?? null;
+        showNotification({
+          title: "Error updating prompt metadata",
+          message,
+          type: "error",
+        });
+      };
+
+      try {
+        const statePrompt = getPrompt(stateRef.current, promptId);
+        if (!statePrompt) {
+          throw new Error(`Could not find prompt with id ${promptId}`);
+        }
+        const prompt = clientPromptToAIConfigPrompt(statePrompt);
+
+        await debouncedUpdatePrompt(
+          prompt.name,
+          {
+            ...prompt,
+            metadata: {
+              ...newMetadata,
+              model: prompt.metadata?.model,
+              parameters: prompt.metadata?.parameters,
+            },
+          },
+          {
+            onError,
+          }
+        );
+      } catch (err: unknown) {
+        onError(err);
+      }
+    },
+    [debouncedUpdatePrompt, showNotification]
+  );
+
   const onUpdatePromptModelSettings = useCallback(
     async (promptId: string, newModelSettings: JSONObject) => {
       if (!debouncedUpdateModel) {
@@ -400,6 +474,7 @@ function AIConfigEditorBase({
         id: promptId,
         modelSettings: newModelSettings,
       });
+      logEventHandler?.("UPDATE_PROMPT_MODEL_SETTINGS");
 
       const onError = (err: unknown) => {
         const message = (err as RequestCallbackError).message ?? null;
@@ -434,7 +509,7 @@ function AIConfigEditorBase({
         onError(err);
       }
     },
-    [debouncedUpdateModel, dispatch, showNotification]
+    [debouncedUpdateModel, dispatch, logEventHandler, showNotification]
   );
 
   const onUpdatePromptModel = useCallback(
@@ -450,6 +525,7 @@ function AIConfigEditorBase({
         id: promptId,
         modelName: newModel,
       });
+      logEventHandler?.("UPDATE_PROMPT_MODEL", { model: newModel });
 
       const onError = (err: unknown) => {
         const message = (err as RequestCallbackError).message ?? null;
@@ -477,7 +553,7 @@ function AIConfigEditorBase({
         onError(err);
       }
     },
-    [dispatch, debouncedUpdateModel, showNotification]
+    [dispatch, debouncedUpdateModel, logEventHandler, showNotification]
   );
 
   const setParametersCallback = callbacks?.setParameters;
@@ -514,6 +590,7 @@ function AIConfigEditorBase({
         type: "UPDATE_GLOBAL_PARAMETERS",
         parameters: newParameters,
       });
+      logEventHandler?.("UPDATE_GLOBAL_PARAMETERS");
 
       const onError = (err: unknown) => {
         const message = (err as RequestCallbackError).message ?? null;
@@ -534,7 +611,7 @@ function AIConfigEditorBase({
         onError(err);
       }
     },
-    [debouncedSetParameters, dispatch, showNotification]
+    [debouncedSetParameters, dispatch, logEventHandler, showNotification]
   );
 
   const onUpdatePromptParameters = useCallback(
@@ -550,6 +627,7 @@ function AIConfigEditorBase({
         id: promptId,
         parameters: newParameters,
       });
+      logEventHandler?.("UPDATE_PROMPT_PARAMETERS");
 
       const onError = (err: unknown) => {
         const message = (err as RequestCallbackError).message ?? null;
@@ -572,7 +650,7 @@ function AIConfigEditorBase({
         onError(err);
       }
     },
-    [debouncedSetParameters, dispatch, showNotification]
+    [debouncedSetParameters, dispatch, logEventHandler, showNotification]
   );
 
   const addPromptCallback = callbacks?.addPrompt;
@@ -646,6 +724,7 @@ function AIConfigEditorBase({
         type: "DELETE_PROMPT",
         id: promptId,
       });
+      logEventHandler?.("DELETE_PROMPT");
 
       try {
         const prompt = getPrompt(stateRef.current, promptId);
@@ -662,7 +741,7 @@ function AIConfigEditorBase({
         });
       }
     },
-    [deletePromptCallback, dispatch, showNotification]
+    [deletePromptCallback, dispatch, logEventHandler, showNotification]
   );
 
   const clearOutputsCallback = callbacks?.clearOutputs;
@@ -676,6 +755,8 @@ function AIConfigEditorBase({
     dispatch({
       type: "CLEAR_OUTPUTS",
     });
+    logEventHandler?.("CLEAR_OUTPUTS");
+
     try {
       await clearOutputsCallback();
     } catch (err: unknown) {
@@ -686,7 +767,7 @@ function AIConfigEditorBase({
         type: "error",
       });
     }
-  }, [clearOutputsCallback, dispatch, showNotification]);
+  }, [clearOutputsCallback, dispatch, logEventHandler, showNotification]);
 
   const runPromptCallback = callbacks?.runPrompt;
 
@@ -842,6 +923,7 @@ function AIConfigEditorBase({
         type: "SET_NAME",
         name,
       });
+      logEventHandler?.("SET_NAME");
 
       await debouncedSetName(name, (err: unknown) => {
         const message = (err as RequestCallbackError).message ?? null;
@@ -852,7 +934,7 @@ function AIConfigEditorBase({
         });
       });
     },
-    [debouncedSetName, showNotification]
+    [debouncedSetName, logEventHandler, showNotification]
   );
 
   const setDescriptionCallback = callbacks?.setConfigDescription;
@@ -885,6 +967,7 @@ function AIConfigEditorBase({
         type: "SET_DESCRIPTION",
         description,
       });
+      logEventHandler?.("SET_DESCRIPTION");
 
       await debouncedSetDescription(description, (err: unknown) => {
         const message = (err as RequestCallbackError).message ?? null;
@@ -895,7 +978,7 @@ function AIConfigEditorBase({
         });
       });
     },
-    [debouncedSetDescription, showNotification]
+    [debouncedSetDescription, logEventHandler, showNotification]
   );
 
   const getState = useCallback(() => stateRef.current, []);
@@ -1003,19 +1086,7 @@ function AIConfigEditorBase({
         <div>
           <Flex justify="flex-end" pt="md" mb="xs">
             {
-              <Group>
-                {downloadCallback && <DownloadButton onDownload={onDownload} />}
-                {shareCallback && <ShareButton onShare={onShare} />}
-                {openInTextEditorCallback && (
-                  <Tooltip label="Open in Text Editor" withArrow>
-                    <ActionIcon
-                      onClick={openInTextEditorCallback}
-                      className="secondaryButton"
-                    >
-                      <IconBraces size="1rem" />
-                    </ActionIcon>
-                  </Tooltip>
-                )}
+              <Group spacing="xs">
                 {!readOnly && onClearOutputs && (
                   <Button
                     loading={undefined}
@@ -1025,6 +1096,32 @@ function AIConfigEditorBase({
                   >
                     Clear Outputs
                   </Button>
+                )}
+                {(downloadCallback || shareCallback) && (
+                  <div>
+                    {downloadCallback && (
+                      <DownloadButton
+                        onDownload={onDownload}
+                        isGrouped={shareCallback != null}
+                      />
+                    )}
+                    {shareCallback && (
+                      <ShareButton
+                        onShare={onShare}
+                        isGrouped={downloadCallback != null}
+                      />
+                    )}
+                  </div>
+                )}
+                {openInTextEditorCallback && (
+                  <Tooltip label="Open in Text Editor" withArrow>
+                    <ActionIcon
+                      onClick={openInTextEditorCallback}
+                      className="secondaryButton"
+                    >
+                      <IconBraces size="1rem" />
+                    </ActionIcon>
+                  </Tooltip>
                 )}
                 {!readOnly && saveCallback && (
                   <Tooltip
@@ -1070,6 +1167,7 @@ function AIConfigEditorBase({
           onChangePromptName={onChangePromptName}
           onDeletePrompt={onDeletePrompt}
           onRunPrompt={onRunPrompt}
+          onUpdatePromptMetadata={onUpdatePromptMetadata}
           onUpdatePromptModel={onUpdatePromptModel}
           onUpdatePromptModelSettings={onUpdatePromptModelSettings}
           onUpdatePromptParameters={onUpdatePromptParameters}