+++ python/src/aiconfig/editor/client/src/components/global/ShareButton.tsx
@@ -1,19 +1,44 @@
-import { Button, Container, Flex, Modal, Text, Tooltip } from "@mantine/core";
+import {
+  Button,
+  Container,
+  Flex,
+  Modal,
+  Text,
+  Tooltip,
+  createStyles,
+} from "@mantine/core";
 import { memo, useContext, useState } from "react";
 import { useDisclosure } from "@mantine/hooks";
 import CopyButton from "../CopyButton";
 import AIConfigContext from "../../contexts/AIConfigContext";
+import { IconShare } from "@tabler/icons-react";
 
 type Props = {
   onShare: () => Promise<string | void>;
+  /**
+   * If grouped, apply styles to the button with assumption that
+   * the button is the rightmost in the group
+   */
+  isGrouped?: boolean;
 };
 
-export default memo(function ShareButton({ onShare }: Props) {
+const useStyles = createStyles(() => ({
+  buttonGroupRight: {
+    borderBottomLeftRadius: 0,
+    borderTopLeftRadius: 0,
+  },
+}));
+
+export default memo(function ShareButton({
+  onShare,
+  isGrouped = false,
+}: Props) {
   const { mode } = useContext(AIConfigContext);
   const [isModalOpened, { open: openModal, close: closeModal }] =
     useDisclosure(false);
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const [shareUrl, setShareUrl] = useState<string>("");
+  const { classes } = useStyles();
 
   const onClick = async () => {
     if (isLoading) {
@@ -42,11 +67,15 @@ export default memo(function ShareButton({ onShare }: Props) {
       <Button
         loaderPosition="center"
         loading={isLoading}
+        loaderProps={{ size: "sm" }}
         onClick={onClick}
         size="xs"
         variant="filled"
+        className={
+          isGrouped ? `${classes.buttonGroupRight} buttonGroupRight` : undefined
+        }
       >
-        Share {artifactName}
+        <IconShare size="20px" />
       </Button>
     </Tooltip>
   );