+++ python/src/aiconfig/editor/client/src/themes/GradioTheme.ts
@@ -24,6 +24,13 @@ export const GRADIO_THEME: MantineThemeOverride = {
 
     return {
       "div.editorBackground": {
+        a: {
+          // Change links back to mantine color instead of gradio override
+          color: `${
+            theme.colorScheme === "light" ? "#1c7ed6" : "#4dabf7"
+          } !important`,
+        },
+
         background: theme.colorScheme === "light" ? "white" : "#0b0f19",
         borderRadius: "8px",
         // Gradio component is iframed so height should be in relation to
@@ -51,6 +58,7 @@ export const GRADIO_THEME: MantineThemeOverride = {
             color: "#ea580c !important",
             fontSize: "16px" /* var(--button-large-text-size) */,
             fontWeight: 600 /* var(--button-large-text-weight) */,
+            minHeight: "36px",
             padding: "0 1.25rem",
           },
 
@@ -88,6 +96,42 @@ export const GRADIO_THEME: MantineThemeOverride = {
           {
             color: "#374151",
           },
+
+        /*
+         * Fix loading spinner color for buttons and loaders rendered in buttons
+         */
+        "button.mantine-Button-root > div.mantine-Button-inner > span.mantine-Button-label > div > svg":
+          {
+            stroke: "#E85921",
+          },
+
+        "button.mantine-Button-root > div.mantine-Button-inner": {
+          "span.mantine-Button-centerLoader > svg": {
+            stroke: "#E85921",
+          },
+        },
+
+        "button.mantine-Button-root[data-loading]::before": {
+          backgroundColor: "rgba(26, 27, 30, 0.2)",
+        },
+
+        ".mantine-Checkbox-root": {
+          ".mantine-Checkbox-input": {
+            borderColor: inputBorderColor,
+
+            "&:checked": {
+              background:
+                "linear-gradient(to bottom right, #ffedd5, #fdba74 100%)",
+            },
+            "&:hover": {
+              background: "linear-gradient(to bottom right, #ffedd5, #fed7aa)",
+            },
+          },
+
+          ".mantine-Checkbox-icon": {
+            color: "#E85921",
+          },
+        },
 
         ".mantine-Input-input:focus": {
           outline: "solid 1px #E85921 !important",
@@ -150,7 +194,7 @@ export const GRADIO_THEME: MantineThemeOverride = {
         ".ghost": {
           input: {
             border: `1px solid ${inputBorderColor}`,
-            maxHeight: "16px",
+            minHeight: "36px",
             fontFamily:
               "sf mono, ui-monospace, Menlo, Monaco, Consolas, 'Liberation Mono', 'Courier New', monospace",
             borderRadius: "8px",
@@ -164,6 +208,16 @@ export const GRADIO_THEME: MantineThemeOverride = {
               backgroundColor: "transparent",
             },
           },
+        },
+
+        ".buttonGroupLeft": {
+          borderBottomRightRadius: 0,
+          borderTopRightRadius: 0,
+        },
+
+        ".buttonGroupRight": {
+          borderBottomLeftRadius: 0,
+          borderTopLeftRadius: 0,
         },
 
         ".cellStyle": {
@@ -186,6 +240,12 @@ export const GRADIO_THEME: MantineThemeOverride = {
             margin: "8px 0px 0px 0px",
             boxShadow: "0px 1px 4px 0px rgba(0, 0, 0, 0.05) inset",
             backgroundColor: inputBackgroundColor,
+          },
+
+          // Override gradio-container ol styles with mantine's
+          ".outputContainer > ol": {
+            margin: "1em 0",
+            paddingLeft: "40px",
           },
         },
 
@@ -217,6 +277,10 @@ export const GRADIO_THEME: MantineThemeOverride = {
               outlineOffset: "-1px",
             },
           },
+
+          ".promptActionBarClosed": {
+            minWidth: "32px",
+          },
         },
 
         ".divider": {
@@ -234,6 +298,11 @@ export const GRADIO_THEME: MantineThemeOverride = {
           margin: "33px 4px 4px 4px",
           padding: "0.625rem !important",
           height: "auto",
+
+          // Make the icon filled when running spinner is shown
+          "div.mantine-Button-inner > span.mantine-Button-label > div > svg": {
+            fill: "#E85921",
+          },
         },
 
         ".runPromptButton.runPromptButtonReadOnly": {